requirements.txt
ipykernel
praw
pandas
beautifulsoup4 
lxml
python-dotenv
huggingface_hub
transformers
peft
datasets
numpy
wandb
evaluate